requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.8.1
atpublic==3.0.1
attrs==21.4.0
backcall==0.2.0
black==22.1.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.6.0
dpath==2.0.6
dulwich==0.20.32
dvc==2.9.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.1.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gcsfs==2022.1.0
gitdb==4.0.9
GitPython==3.1.26
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grandalf==0.6
idna==3.3
imageio==2.16.0
importlib-resources==5.4.0
imutils==0.5.4
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
mailchecker==4.1.11
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nanotime==0.5.2
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.43
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
ply==3.11
prometheus-client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pygit2==1.8.0
Pygments==2.11.2
pygtrie==2.4.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-benedict==0.24.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-fsutil==0.6.0
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rich==11.2.0
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-image==0.19.1
scipy==1.8.0
scmrepo==0.0.7
seaborn==0.11.2
Send2Trash==1.8.0
shortuuid==1.0.8
shtab==1.5.3
six==1.16.0
smmap==5.0.0
stack-data==0.1.4
tabulate==0.8.9
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
tifffile==2022.2.9
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
voluptuous==0.12.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.7.0
